requirements.txt
django<1.9
djangorestframework
psycopg2
django-any-imagefield
pillow
ipython
pytest
django-pytest
